Changed approach of encodings list numeric representation, using a smaller separator of a single digit (with the flexibility to use any separator); other smaller changes, such as an improved dataframe to padded tensor method
Due to the numeric list representation being so big, with such a large separator, there was a loss of information when converting the numeric encodings lists to an integer/string. As such, I tried a different approach, using a much smaller separator, such as a plain 0, while also doing the job of preventing that symbol from appearing in encoding mappings.