Significant optimization of the normalization method; added compatibility with Modin
Turns out that the use of dictionaries to store and retrieve relevant data, like means and standard deviations, was completely unecessary and highly inefficient, to the point that they would break the kernel if the dataframe had many columns. Now the code's execution progress is less trackable, as the normalization isn't done feature by feature, but it's increadibly faster (e.g. from taking ~20 minutes to just ~3 minutes). Still need to address the efficiency in the search of which columns to normalize.

Added options to use Modin, a relatively new Python library that resembles a distributed version of Pandas, similar to Dask but much more easier to implement and more flexible (can handle better very wide dataframes, column wise, not just very long ones, row wise).